+++ babel.config.js
@@ -4,20 +4,6 @@ const presets = [];
 const plugins = [];
 
 switch (env) {
-  case 'cjs':
-    presets.push([
-      '@babel/preset-env',
-      {
-        targets: {
-          node: 'current',
-        },
-        modules: 'commonjs',
-        useBuiltIns: 'usage',
-        corejs: 3,
-      },
-    ]);
-    break;
-
   case 'build':
   default:
     presets.push([
@@ -27,17 +13,10 @@ switch (env) {
           node: 'current',
         },
         modules: false,
-        useBuiltIns: 'usage',
-        corejs: 3,
       },
     ]);
 
-    plugins.push('lodash');
-    plugins.push('./use-lodash-es');
     break;
 }
 
-module.exports = {
-  presets,
-  plugins,
-};
+export default { presets, plugins };