+++ src/Pbjx.js
@@ -1,5 +1,5 @@
-import clamp from 'lodash/clamp';
-import trim from 'lodash/trim';
+import clamp from 'lodash-es/clamp.js';
+import trim from 'lodash-es/trim.js';
 import {
   EVENT_PREFIX,
   SUFFIX_AFTER_HANDLE,
@@ -8,15 +8,15 @@ import {
   SUFFIX_CREATED,
   SUFFIX_ENRICH,
   SUFFIX_VALIDATE,
-} from './constants';
-import getEventNames from './utils/getEventNames';
-import BusExceptionEvent from './events/BusExceptionEvent';
-import GetResponseEvent from './events/GetResponseEvent';
-import InvalidArgumentException from './exceptions/InvalidArgumentException';
-import PbjxEvent from './events/PbjxEvent';
-import RequestHandlingFailed from './exceptions/RequestHandlingFailed';
-import TooMuchRecursion from './exceptions/TooMuchRecursion';
-import ResponseCreatedEvent from './events/ResponseCreatedEvent';
+} from './constants.js';
+import getEventNames from './utils/getEventNames.js';
+import BusExceptionEvent from './events/BusExceptionEvent.js';
+import GetResponseEvent from './events/GetResponseEvent.js';
+import InvalidArgumentException from './exceptions/InvalidArgumentException.js';
+import PbjxEvent from './events/PbjxEvent.js';
+import RequestHandlingFailed from './exceptions/RequestHandlingFailed.js';
+import TooMuchRecursion from './exceptions/TooMuchRecursion.js';
+import ResponseCreatedEvent from './events/ResponseCreatedEvent.js';
 
 const locatorSym = Symbol('locator');
 const maxRecursionSym = Symbol('maxRecursion');
@@ -85,6 +85,7 @@ export default class Pbjx {
    *                               (bind, validate, enrich, etc.)
    * @param {?PbjxEvent} event     An event object containing the message.
    * @param {boolean}    recursive If true, all field values with MessageType are also triggered.
+   * @param {boolean}    throwEx   If true, exceptions are thrown, otherwise they are logged.
    *
    * @returns {Pbjx}
    *
@@ -93,7 +94,7 @@ export default class Pbjx {
    * @throws {TooMuchRecursion}
    * @throws {Exception}
    */
-  async trigger(message, suffix, event = null, recursive = true) {
+  async trigger(message, suffix, event = null, recursive = true, throwEx = false) {
     const fsuffix = `.${trim(suffix, '.')}`;
     if (fsuffix === '.') {
       throw new InvalidArgumentException('Trigger requires a non-empty suffix.');
@@ -122,11 +123,19 @@ export default class Pbjx {
     }
 
     const dispatcher = await this[locatorSym].getDispatcher();
-    await dispatcher.dispatch(`${EVENT_PREFIX}message${fsuffix}`, fevent);
-    const events = getEventNames(message, fsuffix);
+    const events = [`${EVENT_PREFIX}message${fsuffix}`].concat(getEventNames(message, fsuffix));
+
     const l = events.length;
     for (let i = 0; i < l; i += 1) {
-      await dispatcher.dispatch(events[i], fevent);
+      try {
+        await dispatcher.dispatch(events[i], fevent);
+      } catch (e) {
+        if (throwEx) {
+          throw e;
+        }
+        const exceptionHandler = await this[locatorSym].getExceptionHandler();
+        await exceptionHandler.onTriggerException(fevent, events[i], e);
+      }
     }
 
     return this;