+++ src/exceptions/TooMuchRecursion.js
@@ -1,4 +1,4 @@
-import LogicException from './LogicException';
+import LogicException from './LogicException.js';
 
 export default class TooMuchRecursion extends LogicException {
 }