+++ tests/ContainerAwareServiceLocator.test.js
@@ -1,11 +1,11 @@
 import test from 'tape';
-import EchoRequestV1 from '@gdbots/schemas/gdbots/pbjx/request/EchoRequestV1';
-import Container from './fixtures/Container';
-import ContainerAwareServiceLocator from '../src/ContainerAwareServiceLocator';
-import EchoRequestHandler from '../src/EchoRequestHandler';
-import Pbjx from '../src/Pbjx';
-import curieToHandlerServiceId from '../src/utils/curieToHandlerServiceId';
-import { serviceIds } from '../src/constants';
+import EchoRequestV1 from '@gdbots/schemas/gdbots/pbjx/request/EchoRequestV1.js';
+import Container from './fixtures/Container.js';
+import ContainerAwareServiceLocator from '../src/ContainerAwareServiceLocator.js';
+import EchoRequestHandler from '../src/EchoRequestHandler.js';
+import Pbjx from '../src/Pbjx.js';
+import curieToHandlerServiceId from '../src/utils/curieToHandlerServiceId.js';
+import { serviceIds } from '../src/constants.js';
 
 test('ContainerAwareServiceLocator tests', async (t) => {
   const container = new Container();