+++ tests/bootstrap.cjs
@@ -1,0 +1,4 @@
+require('@babel/register')({});
+(async() => {
+  await import('@gdbots/acme-schemas');
+})();